updates for qt

+++ gb_bus.h
@@ -78,7 +78,7 @@ namespace TKPEmu::Gameboy::Devices {
         std::vector<RamBank>& GetRamBanks();
         Cartridge& GetCartridge();
         bool LoadCartridge(std::string filename);
-        std::array<std::array<float, 3>, 4> Palette;
+        std::array<std::array<uint8_t, 3>, 4> Palette;
         std::unordered_map<uint8_t, Change> ScanlineChanges;
         std::array<PaletteColors, 8> BGPalettes{};
         std::array<PaletteColors, 8> OBJPalettes{};
