+++ gb_ppu.h
@@ -20,13 +20,13 @@ namespace TKPEmu::Gameboy::Devices {
 		PPU(Bus& bus, std::mutex* draw_mutex);
 		void Update(uint8_t cycles);
 		void Reset();
-		float* GetScreenData();
+		uint8_t* GetScreenData();
 		void FillTileset(float* pixels, size_t x_off = 0, size_t y_off = 0, uint16_t addr = 0x8000);
 	private:
 		Bus& bus_;
 		std::mutex* draw_mutex_;
-		std::array<float, 4 * 160 * 144> screen_color_data_{};
-		std::array<float, 4 * 160 * 144> screen_color_data_second_{};
+		std::array<uint8_t, 4 * 160 * 144> screen_color_data_{};
+		std::array<uint8_t, 4 * 160 * 144> screen_color_data_second_{};
 		// PPU memory mapped registers
 		uint8_t& LCDC, &STAT, &LYC, &LY, &IF, &SCY, &SCX, &WY, &WX;
 		std::vector<uint8_t> cur_scanline_sprites_;
